+++ spec/libraries/provider_chef_dk_debian_spec.rb
@@ -79,10 +79,4 @@ describe Chef::Provider::ChefDk::Debian do
       p.send(:remove!)
     end
   end
-
-  describe '#bashrc_file' do
-    it 'returns "bash.bashrc"' do
-      expect(provider.send(:bashrc_file)).to eq('/etc/bash.bashrc')
-    end
-  end
 end