+++ spec/support/cookbooks/chef_dk_shell_init_test/recipes/default.rb
@@ -1,0 +1,5 @@
+# Encoding: UTF-8
+
+chef_dk_shell_init 'default' do
+  user node['chef_dk']['shell_init_user']
+end