+++ spec/support/cookbooks/dmg/metadata.rb
@@ -1,9 +1,0 @@
-# Encoding: UTF-8
-
-name             'dmg'
-maintainer       'Jonathan Hartman'
-maintainer_email 'j@p4nt5.com'
-license          'Apache v2.0'
-description      'Installs/configures dmg'
-long_description 'Installs/configures dmg'
-version          '2.2.0'